AI/requirements.txt
python-dotenv
requests
xmltodict
pandas
OpenDartReader
beautifulsoup4
re
langchain
langchain-upstage
langchain-core
langchain-community
pinecone-client
langchain-pinecone
langchain-schema
scikit-learn